+++ HISTORY.md
@@ -1,8 +1,8 @@
 ## Version 4.3.3 - TBA
 
-ENH: Issue #776 TDataSize supports ronna- and quetta- sizes, and uses cultural units for peta- and under (decimal only).
-ENH: Issue #782: CultureInfo loads and formats units from ICU.
-BUG: Issue #1013 CultureInfo loads more than one ICU key.
+ENH: Issue #776 - TDataSize supports ronna- and quetta- sizes, and uses cultural units for peta- and under (decimal only).
+ENH: Issue #782 - CultureInfo loads and formats units from ICU.
+BUG: Issue #1013 - CultureInfo loads more than one ICU key.
 
 ## Version 4.3.2 - January 28, 2026